Backend/requirements.txt
faster-whisper
numpy
scipy
sounddevice
groq
python-dotenv
fastapi
uvicorn
websockets
pipecat-ai[local]
spacy

# HTTP client for Exotel and other provider REST APIs
httpx

# Twilio SDK — optional, only needed for real Twilio calls
# Install with: pip install twilio
# twilio

# System dependency (install separately):
# Ubuntu/Debian: sudo apt install ffmpeg
# macOS:         brew install ffmpeg

# TTS — Edge-TTS (Cloud)
edge-tts
miniaudio

# TTS — Kokoro-82M (Local)
kokoro>=0.9.4
torch
soundfile
# numpy — already listed above
# scipy — already listed above

# System dependency for Kokoro / misaki G2P:
# Ubuntu/Debian: sudo apt install espeak-ng
# macOS:         brew install espeak-ng

sounddevice
